+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Changed
+
+- Differentiate between internal and external file read functions
+
 ## [0.5.3] 2024-03-16
 
 ### Changed